Refactor Editor & Log Editor State
- Editor was getting to more than 600 lines with firestore implementation, so decided to split it off for code clarity
- Can successfully log editorJSON with write button, now needs to be saved appropriately into firebase.
